# quell-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Quell AI Backend (derived from routes)
  version: 1.0.0
  description: |
    OpenAPI specification for the Quell AI backend API. Endpoints cover authentication,
    store management, Shopify OAuth, chat webhook, inbox, support tickets, appearance,
    analytics, and admin metrics.
servers:
  - url: https://histioid-imogene-spinally.ngrok-free.dev
    description: Production
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Store:
      type: object
      properties:
        id: { type: integer }
        shop_domain: { type: string }
        store_id: { type: string }
        widget_token: { type: string }
        product_count: { type: integer }
        status: { type: string }
        last_sync_at: { type: string, format: date-time }
        installed_at: { type: string, format: date-time }
...
security:
  - BearerAuth: []
paths:
...
  /api/stores:
    get:
      tags: [Stores]
      summary: List stores for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Seller stores
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  stores:
                    type: array
                    items: { $ref: "#/components/schemas/Store" }
...
  /admin/platform-overview:
    get:
      tags: [Admin]
      summary: Platform-wide metrics (stores, tokens, tickets)
      parameters:
        - in: query
          name: since
          schema: { type: string, format: date }
      responses:
        '200':
          description: Admin overview
  /admin/top-stores:
    get:
      tags: [Admin]
      summary: Top stores by activity
      parameters:
        - in: query
          name: since
          schema: { type: string, format: date }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: sortBy
          schema: { type: string, enum: [conversations,messages,tokens_used] }
      responses:
        '200':
          description: Top stores list
  /admin/token-usage:
    get:
      tags: [Admin]
      summary: Token usage grouped by store or date
      parameters:
        - in: query
          name: since
          schema: { type: string, format: date }
        - in: query
          name: groupBy
          schema: { type: string, enum: [store,date] }
      responses:
        '200':
          description: Token usage statistics
  /admin/growth:
    get:
      tags: [Admin]
      summary: Store growth over time
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 30 }
      responses:
        '200':
          description: Growth data
  /admin/stores:
    get:
      tags: [Admin]
      summary: List all stores (admin)
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Stores list
  /admin/tickets:
    get:
      tags: [Admin]
      summary: All support tickets across platform
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
      responses:
        '200':
          description: Ticket list
  /admin/stats/daily:
    get:
      tags: [Admin]
      summary: Daily platform statistics
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 30 }
      responses:
        '200':
          description: Daily stats